+++ api/dc/base/serializers.py
@@ -13,7 +13,7 @@ from api.vm.utils import get_vm, get_owners
 from api.sms.utils import get_services
 from api.mon.backends.abstract import VM_KWARGS, VM_KWARGS_NIC, VM_KWARGS_DISK
 from gui.models import User, UserProfile, Role
-from vms.models import Dc, DefaultDc, Vm, BackupDefine, Subnet
+from vms.models import Dc, DefaultDc, Vm, BackupDefine
 from vms.utils import DefAttrDict
 from pdns.models import Domain
 
@@ -203,6 +203,7 @@ class DcSettingsSerializer(s.InstanceSerializer):
         'SMS_SMSAPI_PASSWORD',
     })
     _null_fields_ = frozenset({
+        'VMS_VM_DEFINE_LIMIT',
         'VMS_VM_SNAPSHOT_DEFINE_LIMIT',
         'VMS_VM_SNAPSHOT_LIMIT_AUTO',
         'VMS_VM_SNAPSHOT_LIMIT_MANUAL',
@@ -271,6 +272,10 @@ class DcSettingsSerializer(s.InstanceSerializer):
     VMS_ZONE_ENABLED = s.BooleanField(label='VMS_ZONE_ENABLED',  # Module
                                       help_text=_('Whether to enable support for SunOS and Linux zones in '
                                                   'this virtual datacenter.'))
+
+    VMS_VM_DEFINE_LIMIT = s.IntegerField(label='VMS_VM_DEFINE_LIMIT', required=False,
+                                         help_text=_('Maximum number of virtual servers that can be defined in '
+                                                     'this virtual datacenter.'))
     VMS_VM_STOP_TIMEOUT_DEFAULT = s.IntegerField(label='VMS_VM_STOP_TIMEOUT_DEFAULT',
                                                  help_text='Default time period (in seconds) for a graceful VM stop or '
                                                            'reboot, after which a force stop/reboot is send to the VM '
@@ -411,7 +416,7 @@ class DcSettingsSerializer(s.InstanceSerializer):
                                               help_text=_('Existing Zabbix host group, which will be used for all '
                                                           'monitored servers in this virtual datacenter.'))
     MON_ZABBIX_HOSTGROUPS_VM = s.ArrayField(label='MON_ZABBIX_HOSTGROUPS_VM', max_items=32, required=False,
-                                            help_text=_('List of other existing Zabbix host groups, which will be used '
+                                            help_text=_('List of Zabbix host groups, which will be used '
                                                         'for all monitored servers in this virtual datacenter. '
                                                         'Available placeholders are: {ostype}, {ostype_text}, '
                                                         '{disk_image}, {disk_image_abbr}, {dc_name}.'))
@@ -607,9 +612,6 @@ class DefaultDcSettingsSerializer(DcSettingsSerializer):
     VMS_NODE_SSH_KEYS_DEFAULT = s.ArrayField(label='VMS_NODE_SSH_KEYS_DEFAULT',
                                              help_text=_('List of SSH keys to be added to compute nodes by default'))
 
-    VMS_NET_NIC_TAGS = s.ArrayField(label='VMS_NET_NIC_TAGS', min_items=1, max_items=24,
-                                    help_text=_('List of aliases of network devices configured on compute nodes.'))
-
     VMS_IMAGE_VM = s.SafeCharField(label='VMS_IMAGE_VM', required=False,
                                    help_text=_('Global image server (hostname or uuid) - primary IMGAPI source on all '
                                                'compute nodes. Empty value means that most of the image-related '
@@ -721,21 +723,6 @@ class DefaultDcSettingsSerializer(DcSettingsSerializer):
                         raise s.ValidationError(_('Invalid OS type; VM must be a SunOS Zone.'))
 
                     attrs[source] = vm.uuid
-
-        return attrs
-
-    # noinspection PyMethodMayBeStatic,PyPep8Naming
-    def validate_VMS_NET_NIC_TAGS(self, attrs, source):
-        try:
-            value = attrs[source]
-        except KeyError:
-            pass
-        else:
-            new_nic_tags = set(value)
-            used_nic_tags = set(Subnet.objects.all().values_list('nic_tag', flat=True).distinct())
-
-            if not used_nic_tags.issubset(new_nic_tags):
-                raise s.ValidationError(_('Cannot remove used NIC tags.'))
 
         return attrs