+++ api/image/base/views.py
@@ -159,6 +159,10 @@ image server support is disabled in the system
 
     .. http:delete:: /image/(name)
 
+        .. note:: A server disk image cannot be deleted when it is used by even one virtual server. In order to \
+disable further use of such a disk image, the image can be marked as deleted by \
+:http:put:`changing its access property to deleted (4) </image/(name)>`.
+
         :DC-bound?:
             * |dc-yes| - ``dc_bound=true``
             * |dc-no| - ``dc_bound=false``