+++ api/network/base/serializers.py
@@ -3,20 +3,21 @@ from django.conf import settings
 from frozendict import frozendict
 
 from api import serializers as s
-from api.validators import validate_alias, validate_dc_bound
+from api.validators import validate_alias
 from api.vm.utils import get_owners
 from gui.models import User
-from vms.models import Subnet, IPAddress, DefaultDc
+from vms.models import Subnet, IPAddress, Node
 from pdns.models import Domain
 
 
-class NetworkSerializer(s.InstanceSerializer):
+class NetworkSerializer(s.ConditionalDCBoundSerializer):
     """
     vms.models.Subnet
     """
     _model_ = Subnet
     _update_fields_ = ('alias', 'owner', 'access', 'desc', 'network', 'netmask', 'gateway', 'resolvers',
-                       'dns_domain', 'ptr_domain', 'nic_tag', 'vlan_id', 'dc_bound', 'dhcp_passthrough')
+                       'dns_domain', 'ptr_domain', 'nic_tag', 'vlan_id', 'dc_bound', 'dhcp_passthrough', 'vxlan_id',
+                       'mtu')
     _default_fields_ = ('name', 'alias', 'owner')
     _blank_fields_ = frozenset({'desc', 'dns_domain', 'ptr_domain'})
     _null_fields_ = frozenset({'gateway'})
@@ -32,12 +33,14 @@ class NetworkSerializer(s.InstanceSerializer):
     netmask = s.IPAddressField()
     gateway = s.IPAddressField(required=False)  # can be null
     nic_tag = s.ChoiceField()
+    nic_tag_type = s.CharField(read_only=True)
     vlan_id = s.IntegerField(min_value=0, max_value=4096)
+    vxlan_id = s.IntegerField(min_value=1, max_value=16777215, required=False)  # (2**24 - 1) based on RFC 7348
+    mtu = s.IntegerField(min_value=576, max_value=9000, required=False)  # values from man vmadm
     resolvers = s.IPAddressArrayField(source='resolvers_api', required=False, max_items=8)
     dns_domain = s.RegexField(r'^[A-Za-z0-9][A-Za-z0-9\._-]*$', max_length=250, required=False)  # can be blank
     ptr_domain = s.RegexField(r'^[A-Za-z0-9][A-Za-z0-9\._-]*$', max_length=250, required=False)  # can be blank
     dhcp_passthrough = s.BooleanField(default=False)
-    dc_bound = s.BooleanField(source='dc_bound_bool', default=True)
     created = s.DateTimeField(read_only=True, required=False)
 
     def __init__(self, request, net, *args, **kwargs):
@@ -45,24 +48,13 @@ class NetworkSerializer(s.InstanceSerializer):
         if not kwargs.get('many', False):
             self._dc_bound = net.dc_bound
             self.fields['owner'].queryset = get_owners(request, all=True)
-            self.fields['nic_tag'].choices = [(i, i) for i in DefaultDc().settings.VMS_NET_NIC_TAGS]
+            self.fields['nic_tag'].choices = Node.all_nictags_choices()
 
     def _normalize(self, attr, value):
         if attr == 'dc_bound':
             return self._dc_bound
         # noinspection PyProtectedMember
         return super(NetworkSerializer, self)._normalize(attr, value)
-
-    def validate_dc_bound(self, attrs, source):
-        try:
-            value = bool(attrs[source])
-        except KeyError:
-            pass
-        else:
-            if value != self.object.dc_bound_bool:
-                self._dc_bound = validate_dc_bound(self.request, self.object, value, _('Network'))
-
-        return attrs
 
     def validate_alias(self, attrs, source):
         try:
@@ -121,6 +113,21 @@ class NetworkSerializer(s.InstanceSerializer):
             netmask = self.object.netmask
 
         try:
+            vxlan_id = attrs['vxlan_id']
+        except KeyError:
+            vxlan_id = self.object.vxlan_id
+
+        try:
+            mtu = attrs['mtu']
+        except KeyError:
+            mtu = self.object.mtu
+
+        try:
+            nic_tag = attrs['nic_tag']
+        except KeyError:
+            nic_tag = self.object.nic_tag
+
+        try:
             ip_network = Subnet.get_ip_network(network, netmask)
             if ip_network.is_reserved:
                 raise ValueError
@@ -133,7 +140,31 @@ class NetworkSerializer(s.InstanceSerializer):
 
             if limit is not None:
                 if Subnet.objects.filter(dc_bound=self._dc_bound).count() >= int(limit):
-                    raise s.ValidationError(_('Maximum number of networks reached'))
+                    raise s.ValidationError(_('Maximum number of networks reached.'))
+
+        nic_tag_type = Node.all_nictags()[nic_tag]
+        # retrieve all available nictags and see what is the type of the current nic tag
+        # if type is overlay then vxlan is mandatory argument
+        if nic_tag_type == 'overlay rule':
+            if not vxlan_id:
+                self._errors['vxlan_id'] = s.ErrorList([_('VXLAN ID is required when an '
+                                                          'overlay NIC tag is selected.')])
+        else:
+            attrs['vxlan_id'] = None
+
+        # validate MTU for overlays and etherstubs, and physical nics
+        if nic_tag_type == 'overlay rule':
+            # if MTU was not set for the overlay
+            if not mtu:
+                attrs['mtu'] = 1400
+
+            if mtu > 8900:
+                self._errors['mtu'] = s.ErrorList([s.IntegerField.default_error_messages['max_value']
+                                                   % {'limit_value': 8900}])
+
+        if nic_tag_type in ('normal', 'aggr') and mtu and mtu < 1500:
+            self._errors['mtu'] = s.ErrorList([s.IntegerField.default_error_messages['min_value']
+                                               % {'limit_value': 1500}])
 
         if self._dc_bound:
             try:
@@ -146,7 +177,10 @@ class NetworkSerializer(s.InstanceSerializer):
             if dc_settings.VMS_NET_VLAN_RESTRICT and vlan_id not in dc_settings.VMS_NET_VLAN_ALLOWED:
                 self._errors['vlan_id'] = s.ErrorList([_('VLAN ID is not available in datacenter.')])
 
-        return attrs
+            if dc_settings.VMS_NET_VXLAN_RESTRICT and vxlan_id not in dc_settings.VMS_NET_VXLAN_ALLOWED:
+                self._errors['vxlan_id'] = s.ErrorList([_('VXLAN ID is not available in datacenter.')])
+
+        return super(NetworkSerializer, self).validate(attrs)
 
     # noinspection PyMethodMayBeStatic
     def update_errors(self, fields, err_msg):