+++ gui/dc/network/forms.py
@@ -12,7 +12,7 @@ from api.vm.utils import get_owners
 from gui.forms import SerializerForm
 from gui.fields import ArrayField
 from gui.widgets import ArrayWidget
-from vms.models import Subnet, IPAddress, DefaultDc
+from vms.models import Subnet, IPAddress, Node
 
 TEXT_INPUT_ATTRS = {'class': 'input-transparent narrow', 'required': 'required'}
 SELECT_ATTRS = {'class': 'narrow input-select2'}
@@ -64,6 +64,8 @@ class AdminNetworkForm(SerializerForm):
     nic_tag = forms.ChoiceField(label=_('NIC Tag'), required=True,
                                 help_text=_('NIC tag or device name on compute node.'),
                                 widget=forms.Select(attrs=SELECT_ATTRS))
+    vxlan_id = forms.IntegerField(label=_('VXLAN ID'), required=False, widget=forms.TextInput(attrs=TEXT_INPUT_ATTRS),
+                                  help_text=_('VXLAN ID required for overlay NIC tags (1 - 16777215).'))
 
     # Advanced options
     resolvers = ArrayField(label=_('Resolvers'), required=False,
@@ -80,11 +82,13 @@ class AdminNetworkForm(SerializerForm):
                                           help_text=_('When enabled, IP addresses for this network are managed by '
                                                       'an external DHCP service.'),
                                           widget=forms.CheckboxInput(attrs={'class': 'normal-check'}))
+    mtu = forms.IntegerField(label=_('MTU'), required=False, widget=forms.TextInput(attrs=TEXT_INPUT_ATTRS),
+                             help_text=_('MTU for the network vNIC (576 - 9000)'))
 
     def __init__(self, request, net, *args, **kwargs):
         super(AdminNetworkForm, self).__init__(request, net, *args, **kwargs)
         self.fields['owner'].choices = get_owners(request).values_list('username', 'username')
-        self.fields['nic_tag'].choices = [(i, i) for i in sorted(DefaultDc().settings.VMS_NET_NIC_TAGS)]
+        self.fields['nic_tag'].choices = Node.all_nictags_choices()
 
         if not request.user.is_staff:
             self.fields['dc_bound'].widget.attrs['disabled'] = 'disabled'