+++ vms/models/node.py
@@ -17,6 +17,8 @@ from gui.models import User
 
 NODES_ALL_KEY = 'nodes_list'
 NODES_ALL_EXPIRES = 300
+NICTAGS_ALL_KEY = 'nictag_list'
+NICTAGS_ALL_EXPIRES = None
 
 
 class Node(_StatusModel, _JsonPickleModel, _UserTasksModel):
@@ -406,6 +408,37 @@ class Node(_StatusModel, _JsonPickleModel, _UserTasksModel):
 
         return nodes
 
+    @classmethod
+    def all_nictags(cls, clear_cache=False):
+        """Return dictionary with of nictags {name:type}"""
+        if clear_cache:
+            cache.delete(NICTAGS_ALL_KEY)
+
+        nictags = cache.get(NICTAGS_ALL_KEY)
+
+        if not nictags:
+            nodes = cls.objects.all()
+            nictags = {}
+
+            for node in nodes:
+                for nic in node.nictags:
+                    nic_name = nic['name']
+                    nic_type = nic['type'].replace('aggr', 'normal')
+                    # if nictag name is already present and types are not equal
+                    if nic_name in nictags and nic_type != nictags[nic_name]:
+                        raise ValueError('Duplicate NIC tag name with different type exists on another compute node!')
+
+                    nictags[nic_name] = nic_type
+
+            cache.set(NICTAGS_ALL_KEY, nictags, NICTAGS_ALL_EXPIRES)
+
+        return nictags
+
+    @classmethod
+    def all_nictags_choices(cls):
+        """Return set of tuples that are used as choices in nictag field in NetworkSerializer"""
+        return sorted([(name, '%s (%s)' % (name, typ)) for name, typ in six.iteritems(cls.all_nictags())])
+
     @property
     def _initializing_key(self):
         return 'node:%s:initializing' % self.uuid
@@ -599,8 +632,10 @@ class Node(_StatusModel, _JsonPickleModel, _UserTasksModel):
             if save_ip and self._ip:
                 self._ip.save()
 
-        if clear_cache or status_changed:
-            self.all(clear_cache=True)
+            # this may eventually block the creation or update of the node from sysinfo
+            if clear_cache or status_changed:
+                self.all(clear_cache=True)
+                self.all_nictags(clear_cache=True)
 
         if self.zpool and (zpool_update or zpool_create):
             self.storage.save()  # size parameters were updated during update_resources()
@@ -637,6 +672,7 @@ class Node(_StatusModel, _JsonPickleModel, _UserTasksModel):
         """Clear list of all nodes from cache"""
         ret = super(Node, self).delete(**kwargs)
         self.all(clear_cache=True)
+        self.all_nictags(clear_cache=True)
         return ret
 
     def _get_queue(self, speed):