+++ vms/models/subnet.py
@@ -12,6 +12,7 @@ from vms.mixins import _DcMixin
 # noinspection PyProtectedMember
 from vms.models.base import _VirtModel, _UserTasksModel
 from vms.models.vm import Vm
+from vms.models.node import Node
 
 
 class Subnet(_VirtModel, _DcMixin, _UserTasksModel):
@@ -41,6 +42,8 @@ class Subnet(_VirtModel, _DcMixin, _UserTasksModel):
     ptr_domain = models.CharField(_('PTR Domain'), max_length=255, blank=True,
                                   help_text=_('Name of a Pdns Domain.'))
     dhcp_passthrough = models.BooleanField(_('DHCP Passthrough'), default=False)
+    vxlan_id = models.PositiveIntegerField(_('VXLAN segment ID'), null=True, blank=True, default=None)
+    mtu = models.PositiveIntegerField(_('MTU'), null=True, blank=True, default=None)
 
     class Meta:
         app_label = 'vms'
@@ -96,6 +99,12 @@ class Subnet(_VirtModel, _DcMixin, _UserTasksModel):
     resolvers_api = property(get_resolvers, set_resolvers)
 
     @property
+    def nic_tag_type(self):
+        """Return type of the NIC tag"""
+        # return type of the nictag or empty string if self.nic_tag is not found in Node.all_nictags
+        return Node.all_nictags().get(self.nic_tag, '')
+
+    @property
     def web_data(self):
         """Return dict used in server web templates"""
         return {'dhcp_passthrough': self.dhcp_passthrough}
@@ -114,7 +123,10 @@ class Subnet(_VirtModel, _DcMixin, _UserTasksModel):
             'netmask': self.netmask,
             'gateway': self.gateway,
             'nic_tag': self.nic_tag,
+            'nic_tag_type': self.nic_tag_type,
             'vlan_id': self.vlan_id,
+            'vxlan_id': self.vxlan_id,
+            'mtu': self.mtu,
             'resolvers': self.get_resolvers(),
             'dns_domain': self.dns_domain,
             'ptr_domain': self.ptr_domain,